backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
motor==3.7.1
pymongo==4.14.1
pandas==2.3.2
pillow==11.3.0
pytesseract==0.3.13
opencv-python==4.10.0.84
numpy>=1.26.0,<2.0.0
aiohttp==3.11.11
langchain==0.3.27
langchain-community==0.3.29
langchain-openai==0.3.33
faiss-cpu==1.12.0
openai>=1.104.2,<2.0.0
pydantic-settings==2.10.1
beautifulsoup4==4.12.3
requests==2.32.5
reportlab==4.2.5
